backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
motor==3.6.0
httpx==0.27.2
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
mailjet-rest==1.3.4
python-jose[cryptography]==3.3.0
boto3==1.35.41